refactor(probe): extract shared endpoint-addressing primitives
